feat(theming): tie app logo to active theme
