+++ index.html
@@ -8,10 +8,11 @@
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link
-    href="https://fonts.googleapis.com/css2?family=Exo+2:ital,wght@0,400;0,600;0,800;1,900&family=Inconsolata&family=Work+Sans&display=block"
+    href="https://fonts.googleapis.com/css2?family=Exo+2:ital,wght@0,400;0,600;0,800;1,900&family=Inconsolata&family=M+PLUS+Rounded+1c:wght@300;400;500;700;900&family=Work+Sans&display=block"
     rel="stylesheet">
   <title>OPTN</title>
   <link rel="preload" href="/images/OPTN_logo.png" as="image">
+  <link rel="preload" href="/images/OPTN_logo_sakura.png" as="image">
   <link rel="preload" href="/images/OPTN_header_car.png" as="image">
 </head>
